% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{print.summary.seas_collection}
\alias{print.summary.seas_collection}
\title{Print a \code{summary.seas_collection} object}
\usage{
\method{print}{summary.seas_collection}(x, ...)
}
\arguments{
\item{x}{A \code{summary.seas_collection} object, as returned by
\code{\link[=summary.seas_collection]{summary.seas_collection()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{summary.seas_collection} object
}
