% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{print.summary.seas_sa}
\alias{print.summary.seas_sa}
\title{Print a \code{summary.seas_sa} object}
\usage{
\method{print}{summary.seas_sa}(x, ...)
}
\arguments{
\item{x}{A \code{summary.seas_sa} object, as returned by \code{\link[=summary.seas_sa]{summary.seas_sa()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{summary.seas_sa} object
}
