% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summary.R
\name{print.summary.seas_test}
\alias{print.summary.seas_test}
\title{Print a \code{summary.seas_test} object}
\usage{
\method{print}{summary.seas_test}(x, ...)
}
\arguments{
\item{x}{A \code{summary.seas_test} object, as returned by \code{\link[=summary.seas_test]{summary.seas_test()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{summary.seas_test} object
}
